camera-module/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aioice==0.10.1
aiortc==1.11.0
aiosignal==1.3.2
attrs==25.3.0
av==14.3.0
bidict==0.23.1
cffi==1.17.1
cryptography==44.0.2
dnspython==2.7.0
frozenlist==1.6.0
google-crc32c==1.7.1
h11==0.14.0
idna==3.10
ifaddr==0.2.0
multidict==6.4.3
numpy==2.2.5
opencv-python==4.11.0.86
propcache==0.3.1
pycparser==2.22
pyee==13.0.0
pylibsrtp==0.12.0
pyOpenSSL==25.0.0
python-engineio==4.12.0
python-socketio==5.13.0
simple-websocket==1.1.0
typing_extensions==4.13.2
wsproto==1.2.0
yarl==1.20.0
